Moving block to building.mesh
