+++ semantic/CMakeLists.txt
@@ -23,7 +23,7 @@ set(semantic_SOURCES
   mesh/roof/circular.cpp
   mesh/roof/rectangular.cpp
   mesh/tree.hpp mesh/tree.cpp
-  mesh/block.hpp mesh/block.cpp
+  mesh/multipolymesh.hpp mesh/multipolymesh.cpp
 
   ogr.hpp
   ogr/ogr.cpp ogr/ogr.incl.hpp
