+++ semantic/io.cpp
@@ -156,10 +156,21 @@ void parse(Entity &entity, const Json::Value &value)
     parse(entity.origin, Json::check(value, "origin", Json::arrayValue));
 }
 
+void parse(geometry::MultiPolyMesh<std::string>& m, const Json::Value& value)
+{
+    parse(m.vertices, Json::check(value, "vertices", Json::arrayValue));
+    parse(m.faces, Json::check(value, "faces", Json::arrayValue));
+    parse(m.faceLabels, Json::check(value, "faceLabels", Json::arrayValue));
+}
+
 void parse(Building &building, const Json::Value &value)
 {
     parse(static_cast<Entity&>(building), value);
     parse(building.roofs, Json::check(value, "roofs", Json::arrayValue));
+    if(value.isMember("mesh"))
+    {
+        parse(building.mesh, Json::check(value, "mesh", Json::objectValue));
+    }
 }
 
 void parse(Tree &tree, const Json::Value &value)
@@ -190,19 +201,6 @@ void parse(Railway &railway, const Json::Value &value)
     parse(railway.lines, Json::check(value, "lines", Json::arrayValue));
 }
 
-void parse(geometry::MultiPolyMesh<std::string>& m, const Json::Value& value)
-{
-    parse(m.vertices, Json::check(value, "vertices", Json::arrayValue));
-    parse(m.faces, Json::check(value, "faces", Json::arrayValue));
-    parse(m.faceLabels, Json::check(value, "faceLabels", Json::arrayValue));
-}
-
-void parse(Block &block, const Json::Value &value)
-{
-    parse(static_cast<Entity&>(block), value);
-    parse(block.mesh, Json::check(value, "mesh", Json::objectValue));
-}
-
 template <typename EntityType>
 void parse(std::vector<EntityType> &entities, const Json::Value &container
            , const char *name)
@@ -229,7 +227,6 @@ void parse(World &world, const Json::Value &value)
     parse(world.buildings, value, "buildings");
     parse(world.trees, value, "trees");
     parse(world.railways, value, "railways");
-    parse(world.blocks, value, "blocks");
 }
 
 /* ------------------------------------------------------------------------ */
@@ -349,11 +346,20 @@ void build(Json::Value &value, const Entity &entity
     build(value["origin"], math::Point3(entity.origin + shift));
 }
 
+void build(Json::Value& value,
+           const geometry::MultiPolyMesh<std::string>& mesh)
+{
+    build(value["vertices"], mesh.vertices);
+    build(value["faces"], mesh.faces);
+    build(value["faceLabels"], mesh.faceLabels);
+}
+
 void build(Json::Value &value, const Building &building
            , const math::Point3 &shift)
 {
     build(value, static_cast<const Entity&>(building), shift);
     build(value["roofs"], building.roofs);
+    build(value["mesh"], building.mesh);
 }
 
 void build(Json::Value &value, const Tree &tree
@@ -392,21 +398,6 @@ void build(Json::Value &value, const Railway &railway
     build(value["lines"], railway.lines);
 }
 
-void build(Json::Value& value,
-           const geometry::MultiPolyMesh<std::string>& mesh)
-{
-    build(value["vertices"], mesh.vertices);
-    build(value["faces"], mesh.faces);
-    build(value["faceLabels"], mesh.faceLabels);
-}
-
-void build(Json::Value & value, const Block& block, const math::Point3& shift)
-{ 
-    build(value, static_cast<const Entity&>(block), shift);
-    
-    build(value["mesh"], block.mesh);
-}
-
 template <typename EntityType>
 void build(Json::Value &container, const char *name
            , const std::vector<EntityType> &entities
@@ -430,7 +421,6 @@ void build(Json::Value &value, const World &world)
     build(value, "buildings", world.buildings);
     build(value, "trees", world.trees);
     build(value, "railways", world.railways);
-    build(value, "blocks", world.blocks);
 }
 
 World load(std::istream &is, const fs::path &path)
@@ -568,6 +558,5 @@ void save(const World &world, const fs::path &path
 SEMANTIC_DEFINE_ENTITY_IO_PAIR(Building)
 SEMANTIC_DEFINE_ENTITY_IO_PAIR(Tree)
 SEMANTIC_DEFINE_ENTITY_IO_PAIR(Railway)
-SEMANTIC_DEFINE_ENTITY_IO_PAIR(Block)
 
 } // namespace semantic
