+++ semantic/io.hpp
@@ -74,7 +74,6 @@ struct SerializationOptions : SaveOptions {
 SEMANTIC_DECLARE_ENTITY_IO_PAIR(Building);
 SEMANTIC_DECLARE_ENTITY_IO_PAIR(Tree);
 SEMANTIC_DECLARE_ENTITY_IO_PAIR(Railway);
-SEMANTIC_DECLARE_ENTITY_IO_PAIR(Block);
 
 #undef SEMANTIC_DECLARE_ENTITY_IO_PAIR
