+++ semantic/mesh/mesh.cpp
@@ -28,7 +28,7 @@
 
 #include "../mesh.hpp"
 #include "building.hpp"
-#include "block.hpp"
+#include "multipolymesh.hpp"
 
 namespace semantic {
 
@@ -60,9 +60,24 @@ geometry::Mesh mesh(const Building &building, const MeshConfig &config
     const auto origin(origin_ + building.origin);
 
     geometry::Mesh m;
-    for (const auto &roof : building.roofs) {
-        detail::append(m, mesh(roof, config, origin + roof.center));
+    if (building.mesh.vertices.size())
+    {
+        if (building.roofs.size())
+        {
+            LOG(warn4)
+                << "Building contains both roof(s) and building mesh. Choosing "
+                   "building mesh to produce the semantic mesh.";
+        }
+        detail::append(m, mesh(building.mesh, config, origin));
     }
+    else
+    {
+        for (const auto& roof : building.roofs)
+        {
+            detail::append(m, mesh(roof, config, origin + roof.center));
+        }
+    }
+
     return m;
 }
