+++ semantic/mesh/mesh.incl.hpp
@@ -36,7 +36,6 @@
 #include "roof.hpp"
 #include "building.hpp"
 #include "tree.hpp"
-#include "block.hpp"
 
 namespace semantic {
 
@@ -61,15 +60,6 @@ void mesh(const Tree &tree, const MeshConfig &config
 }
 
 template <typename MeshCallback>
-void mesh(const Block& block,
-          const MeshConfig& config,
-          const math::Point3& origin,
-          const MeshCallback& meshCallback)
-{
-    meshCallback(block, mesh(block, config, origin));
-}
-
-template <typename MeshCallback>
 void mesh(const World &world, const MeshConfig &config
           , const MeshCallback &meshCallback)
 {
@@ -81,7 +71,6 @@ void mesh(const World &world, const MeshConfig &config
     // ENTITY: update when adding a new entity
     SEMANTIC_ENTITY_DISTRIBUTE(buildings);
     SEMANTIC_ENTITY_DISTRIBUTE(trees);
-    SEMANTIC_ENTITY_DISTRIBUTE(blocks);
 
 #undef SEMANTIC_ENTITY_DISTRIBUTE
 }
