+++ semantic/world.hpp
@@ -48,7 +48,6 @@ UTILITY_GENERATE_ENUM(Class,
                       ((building))
                       ((tree))
                       ((railway))
-                      ((block))
                       )
 
 typedef std::vector<Class> Classes;
@@ -75,6 +74,10 @@ struct Building : Entity {
     /** List of roofs.
      */
     roof::Roof::list roofs;
+
+    /** Alternative representation by mesh
+     */
+    geometry::MultiPolyMesh<std::string> mesh;
 };
 
 /** Tree.
@@ -116,18 +119,6 @@ struct Railway : Entity {
     Lines lines;
 };
 
-/** Block subclasses.
- */
-struct Block : Entity
-{
-    /** Entity class.
-     */
-    static const constexpr Class cls = Class::block;
-    typedef std::vector<Block> list;
-
-    geometry::MultiPolyMesh<std::string> mesh;
-};
-
 /** Semantic world.
  *
  * NB: Contains only a list of buildings, so far.
@@ -155,10 +146,6 @@ struct World {
     /** All railways in the world.
      */
     Railway::list railways;
-
-    /** All blocks in the world.
-     */
-    Block::list blocks;
 };
 
 /** Localizes world. Sets world origin center of all world bounding box.
@@ -190,7 +177,6 @@ void distribute(Class cls, const World &world, const Op &op)
     case Class::building: op(world.buildings); break;
     case Class::tree: op(world.trees); break;
     case Class::railway: op(world.railways); break;
-    case Class::block: op(world.blocks); break;
     }
 }
 
@@ -201,7 +187,6 @@ void distribute(Class cls, World &world, const Op &op)
     case Class::building: op(world.buildings); break;
     case Class::tree: op(world.trees); break;
     case Class::railway: op(world.railways); break;
-    case Class::block: op(world.blocks); break;
     }
 }
